Move shape datatypes to dshape
