+++ myia/prim/shape_inferrers.py
@@ -4,6 +4,8 @@ import operator
 from dataclasses import is_dataclass
 from functools import partial, reduce
 
+from ..dshape import NOSHAPE, TupleShape, ListShape, ClassShape, \
+    find_matching_shape
 from ..infer import ANYTHING, GraphInferrer, register_inferrer, \
     PartialInferrer, Track, MyiaShapeError, Inferrer,  MetaGraphInferrer, \
     InferenceError, MyiaTypeError, TransformedReference
@@ -11,7 +13,6 @@ from ..ir import Graph, MetaGraph
 
 from ..dtype import Array, Tuple, List, Class, TypeType, ismyiatype, \
     pytype_to_myiatype
-from ..utils import Named, overload
 
 from . import ops as P
 from .inferrer_utils import static_getter, getelement
@@ -24,72 +25,6 @@ def prod(iterable):
 
 
 shape_inferrer_constructors = {}
-
-
-NOSHAPE = Named('NOSHAPE')
-
-
-class TupleShape:
-    """Class to distinguish the shape of tuples items."""
-
-    __slots__ = ['shape']
-
-    def __init__(self, shape):
-        """Create the shape."""
-        self.shape = tuple(shape)
-
-    def __repr__(self):
-        return f"T{self.shape}"
-
-    def __len__(self):
-        return len(self.shape)
-
-    def __eq__(self, other):
-        return (type(self) == type(other) and
-                self.shape == other.shape)
-
-    def __hash__(self):
-        return hash((type(self), self.shape))
-
-
-class ListShape:
-    """Class to represent the shape of list elements."""
-
-    __slots__ = ['shape']
-
-    def __init__(self, shape):
-        """Create the shape."""
-        self.shape = shape
-
-    def __repr__(self):
-        return f"L{self.shape}"
-
-    def __eq__(self, other):
-        return (type(self) == type(other) and
-                self.shape == other.shape)
-
-    def __hash__(self):
-        return hash((type(self), self.shape))
-
-
-class ClassShape:
-    """Class to represent the shape of dataclass fields."""
-
-    __slots__ = ['shape']
-
-    def __init__(self, shape):
-        """Create the shape."""
-        self.shape = shape
-
-    def __repr__(self):
-        return f"C{self.shape}"
-
-    def __eq__(self, other):
-        return (type(self) == type(other) and
-                self.shape == other.shape)
-
-    def __hash__(self):
-        return hash((type(self), tuple(sorted(self.shape.items()))))
 
 
 class ScalarShapeInferrer(Inferrer):
@@ -106,62 +41,6 @@ class ScalarShapeInferrer(Inferrer):
     def provably_equivalent(self, other):
         """This is always equal to itself."""
         return type(self) == type(other)
-
-
-@overload
-def _generalize_shape(s1: ListShape, s2):
-    if not isinstance(s2, ListShape):
-        raise InferenceError('Mismatched shape types')
-    return ListShape(_generalize_shape(s1.shape, s2.shape))
-
-
-@overload  # noqa: F811
-def _generalize_shape(s1: TupleShape, s2):
-    if not isinstance(s2, TupleShape):
-        raise InferenceError('Mismatched shape types')
-    s1 = s1.shape
-    s2 = s2.shape
-    if len(s1) != len(s2):
-        raise InferenceError('Tuples of different lengths')
-    tup = [_generalize_shape(a, b) for a, b in zip(s1, s2)]
-    return TupleShape(tup)
-
-
-@overload  # noqa: F811
-def _generalize_shape(s1: tuple, s2):
-    if not isinstance(s2, tuple):
-        raise InferenceError('Mismatched shape types')
-    if len(s1) != len(s2):
-        raise InferenceError('Arrays of differing ndim')
-    return tuple(a if a == b else ANYTHING
-                 for a, b in zip(s1, s2))
-
-
-@overload  # noqa: F811
-def _generalize_shape(s1: ClassShape, s2):
-    if not isinstance(s2, ClassShape):
-        raise InferenceError('Mismatched shape types')
-    d = {}
-    if s1.shape.keys() != s2.shape.keys():
-        raise InferenceError('Classes with different fields')
-    for k, v in s1.shape.items():
-        d[k] = _generalize_shape(v, s2.shape[k])
-
-
-@overload  # noqa: F811
-def _generalize_shape(s1: object, s2):
-    if s1 == s2:
-        return s1
-    else:
-        raise InferenceError('Cannot match shapes')
-
-
-def find_matching_shape(shps):
-    """Returns a shape that matches all shapes in `shps`."""
-    s1, *rest = shps
-    for s2 in rest:
-        s1 = _generalize_shape(s1, s2)
-    return s1
 
 
 class ShapeTrack(Track):
